+++ safe_control_gym/controllers/__init__.py
@@ -31,6 +31,10 @@ register(id="mpsc",
          entry_point="safe_control_gym.controllers.mpsc.mpsc:MPSC",
          config_entry_point="safe_control_gym.controllers.mpsc:mpsc.yaml")
 
+register(id="pid",
+         entry_point="safe_control_gym.controllers.pid.pid:PID",
+         config_entry_point="safe_control_gym.controllers.pid:pid.yaml")
+
 register(id="ppo",
          entry_point="safe_control_gym.controllers.ppo.ppo:PPO",
          config_entry_point="safe_control_gym.controllers.ppo:ppo.yaml")
